backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
aiofiles
python-multipart
httpx
python-dotenv
pillow
imagehash
apscheduler
numpy
tensorflow
shapely